requirements-deoldify.txt
# DeOldify + fastai 1.x compatible requirements for Python 3.7/3.8
fastai==1.0.61
torch==1.7.1+cpu
torchvision==0.8.2+cpu
Pillow==6.2.2
numpy==1.19.5
opencv-python
matplotlib
scipy
requests
pyyaml
bottleneck
numexpr
pandas
fastprogress
beautifulsoup4
realesrgan
# NOTE: Install torch and torchvision with the following command for Windows:
# pip install torch==1.7.1+cpu torchvision==0.8.2+cpu -f https://download.pytorch.org/whl/torch_stable.html
